+++ Src/Penneo/(Model)/Document.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Buffers.Text;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -25,18 +26,37 @@ namespace Penneo
             SignType = TYPE_ATTACHMENT;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="cf">Link the document to the case file</param>
         public Document(CaseFile cf)
             : this()
         {
             CaseFile = cf;
         }
 
-        public Document(CaseFile cf, string title, string pdfFilePath)
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="cf">Link the document to the case file</param>
+        /// <param name="title">Add the title of the document as string</param>
+        /// <param name="pdfFile">Add PDF as filepath or base64 encoded file content string</param>
+        public Document(CaseFile cf, string title, string pdfFile)
             : this(cf)
         {
             CaseFile = cf;
             Title = title;
-            PdfFile = pdfFilePath;
+            
+            byte[] buffer = new byte[pdfFile.Length * 3 / 4];
+            if (Convert.TryFromBase64String(pdfFile, buffer, out _))
+            {
+                PdfRaw = buffer;
+            }
+            else if (!string.IsNullOrEmpty(pdfFile) && File.Exists(pdfFile))
+            {
+                PdfFile = pdfFile;
+            }
         }
 
         /// <summary>
@@ -53,7 +73,7 @@ namespace Penneo
         public int? Status { get; set; }
 
         /// <summary>
-        /// Reference to the pdf file on disk, which will be uploaded to the document
+        /// Reference to the pdf file on disk or as a base64 encoded file content string, which will be uploaded to the document
         /// </summary>
         public string PdfFile { get; set; }
 
@@ -70,10 +90,7 @@ namespace Penneo
                 }
                 return _pdfRaw;
             }
-            set
-            {
-                _pdfRaw = value;
-            }
+            set => _pdfRaw = value;
         }
 
         /// <summary>