+++ CHANGELOG.md
@@ -9,6 +9,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.2.0] - 2023-02-09
+### Added
+- Added support for using a base64 string instead of filepath as the document PDF file
+- ```csharp
+  var document = new Document(casefile, "Title", base64AsString);
+- ```
+
 ## [5.1.1] - 2022-11-22
 ### Fixed
 - Fixed an issue that made it impossible to add activation and expiration dates to signature lines